feat: move setup_logging to repo-wide lib dir
