+++ .gitignore
@@ -56,3 +56,4 @@ target/
 tmp
 var/
 venv/
+!src/lib
