+++ src/kubernetes_backup/kubernetes_backup/context.py
@@ -1,0 +1,6 @@
+import sys
+
+from pathlib import Path
+
+path_root = Path(__file__).parents[2]
+sys.path.append(str(path_root))
